refactor(tests): Replace constants of test to mocks folder
