+++ src/components/BottomOfTheBanner/index.test.tsx
@@ -4,9 +4,8 @@ import * as SpinnerModule from "@components/Spinner";
 import * as SwiperModule from "@components/Swiper";
 import * as TodayWeatherModule from "@components/TodayWeather";
 import * as WeatherCardGridModule from "@components/WeatherCardsList";
-import { BOTTOM_BANNER_TEST } from "@constants";
 import { theme } from "@constants/theme";
-import { BOTTOM_OF_THE_BANNER_MOCK } from "@mocks/index";
+import { BOTTOM_BANNER_TEST, BOTTOM_OF_THE_BANNER_MOCK } from "@mocks";
 import { render, screen } from "@testing-library/react";
 import configureMockStore from "redux-mock-store";
 import { ThemeProvider } from "styled-components";
