--- src/constants/elasticSearchTestConstants.ts
@@ -1,12 +1,0 @@
-export const ELASTIC_HOOK_TEST = {
-  DESCRIPTION: "useElastic hook",
-  IT: {
-    RETURNS_INITIAL_VALUES: "should return initial values and functions",
-    FORMATS_CITY_NAME: "should format city name correctly",
-  },
-
-  INITIAL_VALUES: {
-    INPUT_VALUE: "",
-    SHOW_SUGGESTIONS: false,
-  },
-};
