--- src/constants/eventListTestConstants.ts
@@ -1,10 +1,0 @@
-export const EVENT_LIST_TEST = {
-  DESCRIPTION: "EventList Component",
-  IT: {
-    RENDERS_EMPTY_LIST: "renders empty list message when no events",
-    RENDERS_EVENTS_LIST: "renders list of events with dateTime",
-  },
-  CONSTANTS: {
-    EMPTY_LIST_TEXT: "No events at the moment",
-  },
-};
