+++ src/constants/index.ts
@@ -1,32 +1,16 @@
-export * from "./appRouterTestConstants";
-export * from "./bannerTestConstants";
-export * from "./bottomOfTheBannerConstants";
-export * from "./buttonTestConstants";
 export * from "./cacheConstants";
 export * from "./calendarEvents";
 export * from "./conditions";
-export * from "./elasticSearchTestConstants";
 export * from "./envConstants";
 export * from "./errors";
-export * from "./eventListTestConstants";
-export * from "./formatTimeTestConstants";
-export * from "./getCurrentTimeTestConstants";
-export * from "./getErrorMessageTestConstants";
-export * from "./getGeolocationTestConstants";
 export * from "./includesConstants";
-export * from "./isForecastItemTestConstants";
 export * from "./localstorageConstants";
-export * from "./notFoundTestConstants";
 export * from "./numericalConstants";
 export * from "./routes";
-export * from "./spinnerTestConstants";
 export * from "./swiper";
 export * from "./swiperTestConstants";
 export * from "./theme";
 export * from "./time";
-export * from "./timeTestConstants";
-export * from "./transformWeatherTestConstants";
 export * from "./UI";
 export * from "./utilsConstants";
-export * from "./weatherCardListTestConstants";
 export * from "./weatherConstants";
