+++ src/mocks/banner.ts
