--- src/mocks/eventListMock.ts
@@ -1,22 +1,0 @@
-export const EVENT_LIST_MOCKS = {
-  EVENTS: [
-    {
-      id: "1",
-      summary: "Test Event 1",
-      start: { dateTime: "2025-01-01T10:00:00" },
-      end: { dateTime: "2025-01-01T11:00:00" },
-      description: "Description 1",
-    },
-    {
-      id: "2",
-      summary: "Test Event 2",
-      start: { dateTime: "2025-01-02T11:00:00" },
-      end: { dateTime: "2025-01-02T12:00:00" },
-      description: "Description 2",
-    },
-  ],
-  FORMATTED_TIME: {
-    EVENT_1: "10:00",
-    EVENT_2: "11:00",
-  },
-};
