+++ src/mocks/getCurrent.ts
@@ -15,3 +15,23 @@ export const GET_CURRENT_TIME_TEST = {
     },
   },
 };
+
+export const days = ["Sunday", "Monday", "Tuesday", "Wednesday", "Thursday", "Friday", "Saturday"];
+export const months = [
+  "January",
+  "February",
+  "March",
+  "April",
+  "May",
+  "June",
+  "July",
+  "August",
+  "September",
+  "October",
+  "November",
+  "December",
+];
+export const FIRST_DAY_OF_MONTH = 1;
+export const LAST_DAY_OF_MONTH = 31;
+export const SECONDS_IN_MINUTE = 60;
+export const MILLISECONDS_IN_MINUTE = 1000;
