+++ src/mocks/index.ts
@@ -1,6 +1,18 @@
-export * from "./bottomOfTheBannerMocks";
-export * from "./elasticSearchMock";
-export * from "./eventListMock";
-export * from "./geolocationTestMock";
-export * from "./swiperMock";
-export * from "./transformWeatherMock";
+export * from "./appRouter";
+export * from "./banner";
+export * from "./bottomOfTheBanner";
+export * from "./button";
+export * from "./elasticSearch";
+export * from "./eventList";
+export * from "./eventList";
+export * from "./formatTime";
+export * from "./geolocation";
+export * from "./getCurrent";
+export * from "./getErrorMessage";
+export * from "./isForecastItem";
+export * from "./notFound";
+export * from "./spinner";
+export * from "./swiper";
+export * from "./time";
+export * from "./transformWeather";
+export * from "./weatherCardList";
