+++ src/mocks/isForecastItem.ts
