+++ src/mocks/swiper.ts
